pyproject.toml
[project]
name = "newsmgrbot"
version = "0.1.0"
authors = [
    {name = "Mikhail Samylov", email = "134856751+samylovma@users.noreply.github.com"},
]
dependencies = [
    "python-telegram-bot[job-queue,rate-limiter]~=21.4",
    "dishka~=1.3",
    "SQLAlchemy[postgresql-asyncpg]~=2.0",
    "xmltodict~=0.13",
    "uvloop~=0.20",
]
requires-python = "==3.12.*"
readme = "README.rst"

[tool.pdm.dev-dependencies]
dev = [
    "ruff",
    "mypy",
    "types-xmltodict",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203",
    "D213",
    "COM812",
    "ISC001",
    "D1",
    "TRY003",
    "EM101",
]

[tool.mypy]
strict = true
